Move task options config.
